Custom Response type

+++ src/Core/HR.LeaveMangement.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR.LeaveMangement.Application.Exceptions;
 using HR.LeaveMangement.Application.Features.LeaveAllocations.Request.Commands;
 using HR.LeaveMangement.Application.Persistence.Contracts;
+using HR.LeaveMangement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -25,6 +27,12 @@ namespace HR.LeaveMangement.Application.Features.LeaveAllocations.Handlers.Comma
         public async Task Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+
+            if (leaveAllocation == null)
+            {
+                throw new NotFoundException(nameof(leaveAllocation), request.Id);
+
+            }
 
             await _leaveAllocationRepository.Delete(leaveAllocation);
             return;
